requirements.txt
langchain-openai
langchain
python-dotenv
langchain-community
langchain-core
streamlit
pdfplumber
python-docx